[Jisti-Bridge] Move the RoC logic to videoChannel.
Summary: SRTP logic to detecting ROC doesn't quiet work when streams are paused (due to speech activity) and once a stream resumes, it may fail to guess the correct ROC, causing rtp packet drops. Jitsi already had a logic for handing this (sending a few rtp packets every once in a while) but not general enough for our usecase. In this CR, I moved that logic to videoChannel to take advantage of LastNController

Test Plan: tested locally

Reviewers: brian

Subscribers: eng-team-list

Differential Revision: https://phab.fatline.io/D3263
