+++ src/main/java/org/jitsi/videobridge/simulcast/sendmodes/SwitchingSendMode.java
@@ -72,12 +72,6 @@ public class SwitchingSendMode
         = new SimulcastMessagesMapper();
 
     /**
-     * A cyclic counters multitone that counts how many packets we've dropped
-     * per SSRC.
-     */
-    private final CyclicCounters dropped = new CyclicCounters();
-
-    /**
      * The sync root object protecting the access to the simulcast streams.
      */
     private final Object sendStreamsSyncRoot = new Object();
@@ -193,31 +187,6 @@ public class SwitchingSendMode
 
         if (override != null)
             accept = override.matches(pkt);
-
-        if (!accept)
-        {
-            // For SRTP replay protection the webrtc.org implementation uses a
-            // replay database with extended range, using a rollover counter
-            // (ROC) which counts the number of times the RTP sequence number
-            // carried in the RTP packet has rolled over.
-            //
-            // In this way, the ROC extends the 16-bit RTP sequence number to a
-            // 48-bit "SRTP packet index". The ROC is not be explicitly
-            // exchanged between the SRTP endpoints because in all practical
-            // situations a rollover of the RTP sequence number can be detected
-            // unless 2^15 consecutive RTP packets are lost.
-            //
-            // If this variable is set to true, then for every 0x800 (2048)
-            // dropped packets (at most), we send 8 packets so that the
-            // receiving endpoint can update its ROC.
-            //
-            // TODO(gp) We may want to move this code somewhere more centralized
-            // to take into account last-n etc.
-
-            Integer key = pkt.getSSRC();
-            CyclicCounter counter = dropped.getOrCreate(key, 0x800);
-            accept = counter.cyclicallyIncrementAndGet() < 8;
-        }
 
         if (logger.isDebugEnabled())
         {
@@ -675,68 +644,6 @@ public class SwitchingSendMode
                     this.simulcastStreamsChanged(override);
                 }
             }
-        }
-    }
-
-    /**
-     * A thread safe cyclic counter.
-     */
-    static class CyclicCounter
-    {
-        private final AtomicInteger ai = new AtomicInteger(0);
-
-        private final int maxVal;
-
-        public CyclicCounter(int maxVal)
-        {
-            this.maxVal = maxVal;
-        }
-
-        public int cyclicallyIncrementAndGet()
-        {
-            int curVal, newVal;
-            do
-            {
-                curVal = this.ai.get();
-                newVal = (curVal + 1) % this.maxVal;
-                // note that this doesn't guarantee fairness
-            }
-            while (!this.ai.compareAndSet(curVal, newVal));
-            return newVal;
-        }
-    }
-
-    /**
-     * Multitone pattern with Lazy Initialization.
-     */
-    static class CyclicCounters
-    {
-        private final Map<Integer, CyclicCounter> instances
-            = new ConcurrentHashMap<>();
-
-        private final Lock createLock = new ReentrantLock();
-
-        CyclicCounter getOrCreate(Integer key, int maxVal)
-        {
-            CyclicCounter instance = instances.get(key);
-
-            if (instance == null)
-            {
-                createLock.lock();
-                try
-                {
-                    if ((instance = instances.get(key)) == null)
-                    {
-                        instance = new CyclicCounter(maxVal);
-                        instances.put(key, instance);
-                    }
-                }
-                finally
-                {
-                    createLock.unlock();
-                }
-            }
-            return instance;
         }
     }
